Add `Sai::Conversion::RGB::ColorTransformer`
